Rework wheelchairbound
Fixes the save/load bug and fixes it being a singleton.
