+++ Resources/Prototypes/Traits/disabilities.yml
@@ -48,7 +48,8 @@
   name: trait-wheelchair-bound-name
   description: trait-wheelchair-bound-desc
   components:
-    - type: WheelchairBound
+    - type: BuckleOnMapInit
+      prototype: VehicleWheelchair
     - type: LegsParalyzed
 
 - type: trait
